requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sentence-transformers>=2.3.1
qdrant-client>=1.7.3,<1.12
groq>=0.4.2
flashrank>=0.2.0
python-dotenv>=1.0.1
pypdf>=3.17.0
httpx>=0.26.0
tenacity>=8.2.0
langchain-text-splitters>=0.0.1
